requirements.txt
Pillow==10.4.0
streamlit==1.35.0
